Use medium-psql plan, and improve error-handling for the deploy script
Note the change to medium-psql, which enables setting up roles in the DB!
This is necessary for now to ensure that the DB doesn't expose all the tables in the public schema
as writeable. However, this kind of DB can take several minutes to come up! So there's now a loop
in here to wait for that to complete.

A commit to come will remove the parameterization of the anonymous role so that `data/*.sql` files
can be explicit when referring to roles to grant/revoke.